git: Split ntfs and hfs
