+++ hfs_test.go
@@ -1,0 +1,112 @@
+package git
+
+import (
+	"testing"
+
+	"github.com/go-git/go-billy/v6/memfs"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestIsHFSDotGit(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		part string
+		want bool
+	}{
+		{".git", true},
+		{".Git", true},
+		{".GIT", true},
+		{".gIt", true},
+		{".g\u200cit", true},
+		{".gi\u200dt", true},
+		{".gi\ufefft", true},
+		{"\u200e.git", true},
+		{".g\u200ci\u200dt", true},
+		{".gitmodules", false},
+		{".gitignore", false},
+		{".git2", false},
+		{"git", false},
+		{".gxt", false},
+		{"", false},
+		{".", false},
+		{".g\x80it", false},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.part, func(t *testing.T) {
+			t.Parallel()
+			assert.Equal(t, tc.want, isHFSDotGit(tc.part))
+		})
+	}
+}
+
+func TestValidPathProtectHFS(t *testing.T) {
+	t.Parallel()
+
+	fs := newWorktreeFilesystem(memfs.New(), false, true)
+
+	tests := []struct {
+		path    string
+		wantErr bool
+	}{
+		{".git", true},
+		{".g\u200cit", true},
+		{"\u200e.git", true},
+		{".Git", true},
+		{".GIT", true},
+		{".gitignore", false},
+		{"readme.md", false},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.path, func(t *testing.T) {
+			t.Parallel()
+			err := fs.validPath(tc.path)
+			if tc.wantErr {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestValidPathProtectHFSDisabled(t *testing.T) {
+	t.Parallel()
+
+	fs := newWorktreeFilesystem(memfs.New(), false, false)
+
+	hfsPaths := []string{
+		".g\u200cit",
+		"\u200e.git",
+		".gi\ufefft",
+	}
+
+	for _, p := range hfsPaths {
+		t.Run(p, func(t *testing.T) {
+			t.Parallel()
+			err := fs.validPath(p)
+			assert.NoError(t, err, "HFS checks should not apply when protectHFS is false")
+		})
+	}
+}
+
+func TestWorktreeFilesystemRejectsHFSPaths(t *testing.T) {
+	t.Parallel()
+
+	fs := newWorktreeFilesystem(memfs.New(), false, true)
+
+	hfsPaths := []string{
+		".g\u200cit/config",
+		"\u200e.git/config",
+	}
+
+	for _, p := range hfsPaths {
+		t.Run(p, func(t *testing.T) {
+			t.Parallel()
+			_, err := fs.Create(p)
+			assert.Error(t, err, "Create should reject HFS path %q", p)
+		})
+	}
+}
